Updated scss, updated ts files, created types.ts
