+++ src/app/types.ts
@@ -1,0 +1,5 @@
+//interface for list of tasks
+   export interface ITask {
+	task: string;
+	completed: boolean;
+   }
