+++ src/style.scss
@@ -4,6 +4,7 @@ $text_color: #111111;
 $font-family_1: Roboto;
 $basic_background_color: #E8EEFF;
 $disabled_checkbox_color: #9CA59D;
+$checkbox_border_color: #015ADF;
 
 * {
 	box-sizing: border-box;
@@ -12,8 +13,9 @@ $disabled_checkbox_color: #9CA59D;
 }
 .content {
 	position: relative;
-	height: 100vh;
-	margin: auto;
+	width: 650px;
+    min-height: 100vh;
+	margin: 0 auto;
 	padding: 40px;
 	background-color: $basic_background_color;
 	display: flex;
@@ -39,28 +41,28 @@ h1 {
 	position: absolute;
 	&:checked + .check__box {
 				background-color: $basic_color;
-				background-image: url("../src/assets/images/check.svg");
+				background-image: url("assets/images/check.svg");
 				background-repeat: no-repeat;
 				background-position: center;
 			}
 		&:disabled + .check__box {
 					background-color: $disabled_checkbox_color;
-					background-image: url("../src/assets/images/check.svg");
+					background-image: url("assets/images/check.svg");
 					background-repeat: no-repeat;
 					background-position: center;
 				}
 	&:focus + .check__box {
-				box-shadow: 0 0 0 2px #5C9DFE, 0 0 0 4px #015ADF;
+				box-shadow: 0 0 0 2px $basic_color, 0 0 0 4px $checkbox_border_color;
 			}
 	&:disabled + .check__box {
-				box-shadow: 0 0 0 2px #9CA59D;
+				box-shadow: 0 0 0 2px $disabled_checkbox_color;
 			}
 }
 .check__box {
 	position: absolute;
 	width: 24px;
 	height: 24px;
-	box-shadow: 0 0 0 2px #5C9DFE;
+	box-shadow: 0 0 0 2px $basic_color;
     left: 0;
 	border-radius: 6px;
 	background-color: $basic_background_color;
@@ -82,6 +84,6 @@ h1 {
 	right: 0;
 	width: 50px;
 	height: 50px;
-	background-image: url("../src/assets/images/add-task.svg");
+	background-image: url("assets/images/add-task.svg");
 	background-repeat: no-repeat;
 }
